changed the instance variable id from Integer to a Long. Updated getters and setters. Went to UserService and created an Optional<User> datatype to find users by id. Went to UserServiceImpl, changed the COUNTER from integer to Long and added an l at end of number because Long complains. Created a Optional method in ServiceImpl using stream and filter from class to grab the getter getId and properly set up method. Went to UserController and learned about new annotations like GetMapping with a new / PATH and PathVariable that is needed to properly communicate with the instance variable and the http request. Wrote a similar method in controller about finding the id and needed an if statement. Set up and now works

+++ Server/demo/src/main/java/com/model/User.java
@@ -9,7 +9,7 @@ package com.model;
 
 public class User {
 	
-	private Integer id; 
+	private Long id; 
 	private String firstName;
 	private String lastName;
 	private Integer age;
@@ -21,7 +21,7 @@ public class User {
 	}
 	
 
-	public User(Integer id, String firstName, String lastName, Integer age, String country) {
+	public User(Long id, String firstName, String lastName, Integer age, String country) {
 		super();
 		this.id = id; 
 		this.firstName = firstName;
@@ -31,12 +31,12 @@ public class User {
 	}
 
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
